transfer all commands to class cmd
